Setup 1.20.1 project, WIP migrating dynamic registries to datagen

+++ src/main/java/com/github/mixinors/astromine/common/block/entity/utility/PumpBlockEntity.java
@@ -165,7 +165,7 @@ public class PumpBlockEntity extends ExtendedBlockEntity implements FluidStorage
 							var posesChecked = new LongOpenHashSet();
 							var posToCheck = new ArrayDeque<BlockPos>();
 							
-							posToCheck.add(getPos().add(0, -Math.ceil(depth / 20.0D), 0));
+							posToCheck.add(getPos().add(0, (int) -Math.ceil(depth / 20.0D), 0));
 							
 							var mainCheckFluidState = world.getFluidState(posToCheck.getLast());
